Refactored DimAuto unit tests
